requirements.txt
Django==4.2.30
Flask==3.1.3
Jinja2==3.1.6
MarkupSafe==3.0.3
Werkzeug==3.1.8
amqp==5.3.1
annotated-types==0.7.0
anthropic==0.97.0
anyio==4.12.1
asgiref==3.11.1
async-timeout==5.0.1
billiard==4.2.4
blinker==1.9.0
celery==5.6.3
certifi==2026.4.22
charset-normalizer==3.4.7
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
click==8.1.8
distro==1.9.0
dj-database-url==2.2.0
djangorestframework==3.16.1
docstring_parser==0.18.0
dukpy==0.5.1
esprima==4.0.1
exceptiongroup==1.3.1
greenlet==3.2.5
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.13
importlib_metadata==8.7.1
itsdangerous==2.2.0
jiter==0.14.0
kombu==5.6.2
nodeenv==1.10.0
openai==2.32.0
packaging==26.2
playwright==1.59.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.12
pydantic==2.13.3
pydantic_core==2.46.3
pyee==13.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
redis==7.0.1
requests==2.32.5
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
stripe==15.1.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.2
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
whitenoise==6.7.0
zipp==3.23.1
